testcases/test_framework.py
"""
框架验证测试用例
用于验证测试框架是否正常工作，不依赖浏览器
"""
import allure
import pytest


@allure.feature("框架验证")
@allure.story("基础功能验证")
class TestFrameworkValidation:
    """
    框架验证测试类
    验证 pytest、allure、配置加载等基础功能
    """

    @allure.title("pytest 基础功能验证")
    @allure.description("验证 pytest 基础断言功能正常")
    @pytest.mark.smoke
    def test_pytest_basic(self):
        """测试 pytest 基础功能"""
        assert 1 + 1 == 2
        assert "hello" in "hello world"

    @allure.title("配置文件加载验证")
    @allure.description("验证配置文件可以正常加载")
    @pytest.mark.smoke
    def test_config_loading(self, config):
        """测试配置加载"""
        assert config is not None
        assert "base_url" in config
        assert "browser" in config

    @allure.title("测试数据加载验证")
    @allure.description("验证测试数据可以正常加载")
    @pytest.mark.smoke
    def test_test_data_loading(self, test_data):
        """测试数据加载"""
        assert test_data is not None

    @allure.title("管理员账号数据验证")
    @allure.description("验证管理员账号配置正确（三字段登录）")
    @pytest.mark.smoke
    def test_admin_credentials(self, admin_credentials):
        """测试管理员账号数据"""
        assert "org_code" in admin_credentials
        assert "phone" in admin_credentials
        assert "password" in admin_credentials
        assert admin_credentials["org_code"] != ""
        assert admin_credentials["phone"] != ""
        assert admin_credentials["password"] != ""

    @allure.title("Allure 附件功能验证")
    @allure.description("验证 Allure 附件功能正常")
    @pytest.mark.smoke
    def test_allure_attachment(self):
        """测试 Allure 附件功能"""
        allure.attach(
            "这是一段测试文本",
            name="测试附件",
            attachment_type=allure.attachment_type.TEXT
        )
        assert True

    @allure.title("Allure 步骤功能验证")
    @allure.description("验证 Allure 步骤功能正常")
    @pytest.mark.smoke
    def test_allure_steps(self):
        """测试 Allure 步骤功能"""
        with allure.step("步骤1: 执行操作A"):
            result_a = True

        with allure.step("步骤2: 执行操作B"):
            result_b = True

        with allure.step("步骤3: 验证结果"):
            assert result_a and result_b