requirements.txt
pytest==8.1.1
selenium==4.19.0
allure-pytest==2.13.5
webdriver-manager==4.0.1
PyYAML==6.0.1
openpyxl==3.1.2
pandas==2.2.1
requests==2.31.0
pytest-html==4.1.1
pytest-xdist==3.5.0
pytest-rerunfailures==14.0
Pillow==10.3.0
